requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-modeltranslation==0.19.3
djangorestframework==3.15.1
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.8
packaging==24.0
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.5
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
uritemplate==4.1.1
zope.interface==6.4.post2
